Pacer.Command.Chart

Description

Chart functionality.

Top level command

handle :: forall (es :: [Effect]). (HasCallStack, FileReader :> es, FileWriter :> es, Logger :> es, LoggerNS :> es, PathReader :> es, PathWriter :> es, TypedProcess :> es) => ChartParamsFinal -> Eff es ()

Handles chart command.

Low level functions

createCharts :: forall (es :: [Effect]). (HasCallStack, FileReader :> es, FileWriter :> es, Logger :> es, LoggerNS :> es, PathReader :> es, PathWriter :> es, TypedProcess :> es) => ChartParamsFinal -> Eff es ()

createChartsJsonFile

Arguments

:: forall (es :: [Effect]). (HasCallStack, FileReader :> es, FileWriter :> es, Logger :> es, LoggerNS :> es, PathWriter :> es)

=> Bool
Level to log the success message at.

-> Maybe RunsType

-> Path Abs File

-> Path Abs File

-> Path Abs File

-> Eff es ()

Given ChartParamsFinal, generates a json-encoded array of charts, and writes the file to the given location.

createChartsJsonBS

Arguments

:: forall (es :: [Effect]). (HasCallStack, FileReader :> es, Logger :> es, LoggerNS :> es)

=> Maybe RunsType

-> Path Abs File
Path to runs file. Defaults to defRunsPath.

-> Path Abs File
Path to chart-requests.toml. Defaults to defChartRequestsPath.

-> Eff es LazyByteString

Given file paths to runs and chart requests, returns a lazy json-encoded bytestring of a chart array.