Pacer.Command.Chart.Data.Activity

Activity

data Activity (dist :: DistanceUnit) a

Type for activities.

Constructors

MkActivity

Fields

atype :: Maybe ActivityType

The type of the activity. This is only used in filtering, hence it is arguably redundant due to the existence of labels. Its only justification is for interfacing with garmin (which has exactly one type), and though we could just parse that into labels, it is probably simpler to separate them.

datetime :: Timestamp

The start time of the activity.

distance :: Distance dist a

The activity's total distance.

duration :: Duration a

The activity's total duration.

labels :: Set Label

Optional labels.

title :: Maybe Text

Optional title.

newtype ActivityType

Activity type.

Constructors

MkActivityType

Fields

unActivityType :: Text

newtype Label

Activity label.

Constructors

MkLabel

Fields

unLabel :: Text

Functions

derivePace :: forall a (d :: DistanceUnit). (MGroup a, PaceDistF d) => Activity d a -> Pace d a

Derives the pace from a activity.

SomeActivity

data SomeActivity a where

Existential quantifies an Activity's distance.

Constructors

MkSomeActivity :: forall (d :: DistanceUnit) a. Sing d -> Activity d a -> SomeActivity a

someActivityIso :: forall (d :: DistanceUnit) a. (Fromℤ a, Ord a, Semifield a, Show a, SingI d) => Iso' (SomeActivity a) (Activity d a)

Iso between SomeActivity and Activity. Note that this converts the underlying distance to the requested unit, hence the actual value _can_ change. Thus the isomorphism must be understood in terms of the hypothetical equivalence class on distance units.

someActivityApplyActivity :: (forall (d :: DistanceUnit). Activity d a -> b) -> SomeActivity a -> b

Applies the function to the underlying activity.

Functions

deriveSomePace :: (Fromℤ a, Ord a, Semifield a, Show a) => SomeActivity a -> SomePace a

Derives a pace from some activity.

SomeActivities

newtype SomeActivitiesKey a

Key for SomeActivities. Eq/Ord use an equivalence class on the timestamp, used to enforce that timestamps are unique.

Constructors

MkSomeActivitiesKey

Fields

unSomeActivitiesKey :: SomeActivity a

someActivitiesKeyIso :: Iso' (SomeActivitiesKey a) (SomeActivity a)

Iso between SomeActivitiesKey and SomeActivity.

newtype SomeActivities a

Holds multiple activities.

Constructors

MkSomeActivities

Fields

unSomeActivities :: NESet (SomeActivitiesKey a)

Construction

mkSomeActivitiesFail :: MonadFail m => [SomeActivity a] -> m (SomeActivities a)

mkSomeActivities :: NonEmpty (SomeActivity a) -> Result ActivityDatetimeOverlapE (SomeActivities a)

data ActivityDatetimeOverlapE

Constructors

MkActivityDatetimeOverlapE

Fields

r1 :: (Maybe Text, Timestamp)

r2 :: (Maybe Text, Timestamp)

Elimination

someActivitiesToNE :: SomeActivities a -> NonEmpty (SomeActivity a)

someActivitiesToList :: SomeActivities a -> [SomeActivity a]

Functions

mapSomeActivities :: (forall (d :: DistanceUnit). Activity d a -> Activity d b) -> SomeActivities a -> SomeActivities b

mapAccumSomeActivities :: forall a1 a2 b. Semigroup b => (forall (d :: DistanceUnit). Activity d a1 -> (Activity d a2, b)) -> SomeActivities a1 -> (SomeActivities a2, b)

unionSomeActivities :: SomeActivities a -> SomeActivities a -> Result ActivityDatetimeOverlapE (SomeActivities a)