Pacer.Command.Chart.Data.Chart

Documentation

data Chart

Holds all data associated to a single chart.

Constructors

MkChart

Fields

chartData :: ChartData

Chart data.

chartExtra :: ChartExtra

Chart extra.

chartOptions :: ChartOptions

Chart options.

mkChart :: (Fromℤ a, MetricSpace a, Ord a, Semifield a, Show a, Toℝ a) => SomeRuns a -> ChartRequest a -> Either CreateChartE Chart

mkCharts :: (Fromℤ a, MetricSpace a, Ord a, Semifield a, Show a, Toℝ a) => SomeRuns a -> ChartRequests a -> Either CreateChartE (Seq Chart)

Given runs and chart requests, generates a series of charts, or the