Pacer.Command.Chart.Data.ChartData

Documentation

data ChartData

Holds chart data.

Constructors

ChartDataY ChartY

ChartDataY1 ChartY1

data ChartY

Data for a chart with a single Y axis.

Constructors

MkChartY

Fields

values :: NESeq (Timestamp, Double)

X and Y axis data.

yType :: YAxisType

Y axis type. This is used for the label on the line itself, not the y-axis (that label is on ChartOptions).

yDistUnit :: DistanceUnit

Distance unit for y-axis, if it used.

data ChartY1

Data for a chart with two Y axes.

Constructors

MkChartY1

Fields

values :: NESeq (Timestamp, Double, Double)

Data for a chart with two y Axes.

yType :: YAxisType

Type of first Y axis.

y1Type :: YAxisType

Type of second Y axis.

yDistUnit :: DistanceUnit

Distance unit for y-axis, if it used.

mkChartData

Arguments

:: forall (es :: [Effect]) a. (Display a, Fromℤ a, MetricSpace a, Logger :> es, Ord a, Reader LogEnv :> es, Semifield a, Show a, Toℝ a)

=> DistanceUnit
Final distance unit to use.

-> [FilterExpr a]
Global filters

-> SomeActivities a
List of activities.

-> ChartRequest a
Chart request.

-> Eff es (Result CreateChartE ChartData)
ChartData result. Nothing if no activities passed the request's filter.

Turns a sequence of activities and a chart request into a chart.