Pacer.Command.Chart.Data.Expr

Expressions

data Expr a

Expressions upon which we can filter

Constructors

Atom a
"expr"

Not (Expr a)
"not (expr)"

Or (Expr a) (Expr a)
"or (expr) (expr)"

And (Expr a) (Expr a)
"and (expr) (expr)"

type FilterExpr a = Expr (FilterType a)

Alias for a filter expression.

eval :: (a -> Bool) -> Expr a -> Bool

Parsing

data ExprToken

Constructors

TokenString Text

TokenExprAnd

TokenExprNot

TokenExprOr

TokenExprXor

TokenParenL

TokenParenR

exprLexer :: MParser [ExprToken]

exprParser :: Parsec Void [ExprToken] (Expr Text)

Filtering

data FilterType a

Ways in which we can filter runs.

Constructors

FilterDistance FilterOp (SomeDistance a)

FilterDuration FilterOp (Duration a)

FilterLabel Text
Filters by label equality.

FilterDate FilterOp Moment

FilterPace FilterOp (SomePace a)

data FilterOp

Operator for filter comparisons. The text field is just so we can have Eq and Show instances, though they are of arguably little value. data FilterOp = MkFilterOp Text (forall a. (Ord a) => a -> a -> Bool)

Constructors

FilterOpEq

FilterOpNeq

FilterOpLte

FilterOpLt

FilterOpGte

FilterOpGt

Misc

lexParse :: Parser a => Text -> ResultDefault (Expr a)