Pacer.Command.Chart.Data.Expr.Set

Description

Provides filters over generic set.

FilterElem

data FilterElem (p :: Symbol) a

FilterSet X tests set operations on some element x e.g.

x = y

x ∈ Y

X can be a single value (e.g. type) or a set itself (e.g. labels set, written as singular 'label') e.g.

"type = some_type"

"type ∈ {t1, t2}"

When X is a set, the element x should be understood to be an arbitrary x ∈ X i.e. we are dealing with existence:

∃x ∈ X s.t. x = y

∃x ∈ X s.t. x ∈ Y

E.g.

"label = some_label" ≣ ∃l ∈ activity.labels s.t. l = "some_label"

"label ∈ {l1, l2}"   ≣ ∃l ∈ activity.labels s.t. l ∈ {"l1", "l2"}

The second allows us to encode disjunction (OR) at the operator level. Note that negation applies to the _entire_ expression i.e. the above becomes:

∄x ∈ X s.t. x = y ≣ ∀x ∈ X, x ≠ y

∄x ∈ X s.t. x ∈ Y ≣ ∀x ∈ X, x ∉ Y

E.g.

"label ≠ some_label" ≣ ∀l ∈ activity.labels, l ≠ "some_label"

"label ∉ {l1, l2}"   ≣ ∀l ∈ activity.labels, l ∉ {"l1", "l2"}

Constructors

FilterElemEq FilterOpEq a
Equality test.

FilterElemExists FilterElemOpSet (Set a)
Membership test.

data FilterElemOpSet

Operator for LHS element and RHS set.

Constructors

FilterElemInSet
Membership.

FilterElemNotInSet
Not membership.

memberFun :: Ord a => FilterElemOpSet -> a -> Set a -> Bool

FilterSet

data FilterSet (p :: Symbol) a

FilterSet X tests set operations on the set X e.g. X = "labels". Note that, wrt the operators, the LHS is always a set (e.g. labels set), so operators where the LHS is expected to be an element -- FilterOpEq used by FilterSetExistsElem -- the LHS is considered to be an arbitrary element in the set.

For instance, label = foo means "there exists an l in labels s.t. l == foo".

Constructors

FilterSetElem (FilterElem p a)
Operations on an arbitrary element in X. Not only does this allow alternative syntax for X ∋ a -- i.e. x = a, it also allows encoding OR e.g. x ∈ {a, b} for any x in X.

FilterSetHasElem FilterSetOpElem a
Set membership.

FilterSetComp FilterSetOpSet (Set a)
Set comparisons e.g. labels ⊇ {label_1, label_2}. This tests that all labels exist within labels (AND).

data FilterSetOpElem

Operatoror for LHS set and RHS element.

Constructors

FilterSetContainsElem
Membership.

FilterSetNContainsElem
Not membership.

data FilterSetOpSet

Operator for set comparisons.

Constructors

FilterSetOpSetEq FilterOpEq
Activity set X must equal the given set.

FilterSetOpSetPSuper
Activity set X must be a proper superset of the given set.

FilterSetOpSetSuper
Activity set X must be a superset of the given set.

FilterSetOpSetPSub
Activity set X must be a proper subset of the given set.

FilterSetOpSetSub
Activity set X must be a subset of given set.

existsElemFun :: Ord a => FilterOpEq -> Set a -> a -> Bool

existsSetFun :: Ord a => FilterElemOpSet -> Set a -> Set a -> Bool

hasElemFun :: Ord a => FilterSetOpElem -> Set a -> a -> Bool

compFun :: Ord b => FilterSetOpSet -> Set b -> Set b -> Bool

Misc

parseTextNonEmpty :: MParser Text