Pacer.Command.Chart.Data.Time.Timestamp

Timestamp

data Timestamp

Timestamp for activities.

Construction

fromLocalTime :: MonadFail m => LocalTime -> m Timestamp

Elimination

timestampToYear :: MonadFail m => Timestamp -> m Year

timestampToYearMonth :: MonadFail m => Timestamp -> m (Year, Month)

timestampToDay :: Timestamp -> Day

Predicates

sameYear :: HasCallStack => Timestamp -> Timestamp -> Bool

sameMonth :: HasCallStack => Timestamp -> Timestamp -> Bool

sameWeek :: Timestamp -> Timestamp -> Bool

sameInterval

Arguments

:: Day
Start day d

-> Word16
Interval period (i.e. number of days) p

-> Timestamp
ts1

-> Timestamp
ts2

-> Bool

Determines if two timestamps belong to the same interval vi where vi is an element in d, d + p, d + 2p, .... In other words, each timestamp will belong in some interval d + pk. sameInterval determines if they are the same.

Rounding

roundTimestampYear :: Timestamp -> Timestamp

roundTimestampMonth :: Timestamp -> Timestamp

roundTimestampWeek :: Timestamp -> Timestamp

roundInterval :: Day -> Word16 -> Timestamp -> Timestamp

Rounds the timestamp down to the start of its interval. That is, finds the interval containing the given timestamp, and returns the interval's start day.

Formatting

fmtTimestamp :: Timestamp -> Text

Misc

strictOverlaps :: Timestamp -> [Timestamp]

Returns a list of timestamps that overlap with this one. Strict in the sense that the timestamp itself is not included.