Pacer.Command.Chart.Params

Description

Chart parameters.

Types

data ChartParams (p :: ConfigPhase)

Chart params.

Constructors

MkChartParams

Fields

activityLabelsPath :: MPathF p File

Optional path to activity labels file.

activityPaths :: ActivityPathsF p

Optional path to activities file.

buildDir :: BuildDirF p

Build directory.

cleanInstall :: Bool

If true, copies clean install of web dir.

chartRequestsPath :: PathF p File

Optional path to chart-requests.json.

dataDir :: PathF p Dir

Optional path to directory with activities file and chart-requests.json.

json :: Bool

If true, stops the build after generating the intermediate json file.

port :: ConfigPhaseF p Word16

Port the server runs on.

type ChartParamsArgs = ChartParams 'ConfigPhaseArgs

type ChartParamsFinal = ChartParams 'ConfigPhaseFinal

data ActivitiesType

The type of activities file.

Constructors

ActivitiesDefault
Default activities type, activities.json.

ActivitiesGarmin
Garmin activities type, activities.csv.

Functions

evolvePhase :: forall (es :: [Effect]). (HasCallStack, Logger :> es, LoggerNS :> es, PathReader :> es, Reader LogEnv :> es, State CachedPaths :> es) => ChartParamsArgs -> Maybe ConfigWithPath -> Eff es ChartParamsFinal

Evolve chart params' phase.

Type families

type family ActivityPathsF (p :: ConfigPhase) where ...

Equations

ActivityPathsF 'ConfigPhaseArgs = Seq OsPath

ActivityPathsF 'ConfigPhaseFinal = NESeq (Path Abs File)

type family BuildDirF (p :: ConfigPhase) where ...

Equations

BuildDirF 'ConfigPhaseArgs = Maybe OsPath

BuildDirF 'ConfigPhaseFinal = Path Abs Dir

type family MPathF (p :: ConfigPhase) t where ...

Equations

MPathF 'ConfigPhaseArgs _1 = Maybe OsPath

MPathF 'ConfigPhaseFinal File = Maybe (Path Abs File)

MPathF 'ConfigPhaseFinal Dir = ()

MPathF 'ConfigPhaseFinal t = TypeError (('Text "Type '" ':<>: 'ShowType t) ':<>: 'Text "' invalid for MPathF.") :: Type

type family PathF (p :: ConfigPhase) t where ...

Equations

PathF 'ConfigPhaseArgs _1 = Maybe OsPath

PathF 'ConfigPhaseFinal File = Path Abs File

PathF 'ConfigPhaseFinal Dir = ()

PathF 'ConfigPhaseFinal t = TypeError (('Text "Type '" ':<>: 'ShowType t) ':<>: 'Text "' invalid for PathF.") :: Type