Pacer.Command.Derive

Description

Derive functionality.

Top level handler

handle :: forall (es :: [Effect]) a. (Display a, Fromℤ a, HasCallStack, Ord a, Semifield a, Show a, Terminal :> es, Toℚ a) => DeriveParamsFinal a -> Eff es ()

Handles derive command.

Low level functions

Distance

deriveDistance

Arguments

:: forall (d :: DistanceUnit) a. MGroup a

=> Duration a
Distance.

-> Pace d a
Pace.

-> Distance d a
Distance.

Given a duration and pace, derives the distance.

deriveSomeDistance

Arguments

:: MGroup a

=> Duration a
Duration.

-> SomePace a
Existentially-quantified Pace.

-> SomeDistance a
Existentially-quantified Distance.

Given a duration and existentially-quantified pace, derives the distance.

Duration

deriveDuration

Arguments

:: forall a (d :: DistanceUnit). MSemigroup a

=> Distance d a
Distance.

-> Pace d a
Pace.

-> Duration a
Duration.

Given a distance and pace, derives the duration.

deriveSomeDuration

Arguments

:: (Fromℤ a, Ord a, Semifield a, Show a)

=> SomeDistance a
Existentially-quantified distance.

-> SomePace a
Existentially-quantified pace.

-> Duration a
Duration.

Given existentially-quantified distance and pace, derives the duration. Different distance units are converted.

Pace

derivePace

Arguments

:: forall (d :: DistanceUnit) a. (MGroup a, PaceDistF d)

=> Distance d a
Distance.

-> Duration a
Duration.

-> Pace d a
Pace.

Given a distance and a duration, derives the pace.

deriveSomePace

Arguments

:: (Fromℤ a, Ord a, Semifield a, Show a)

=> SomeDistance a
Existentially-quantified distance.

-> Duration a
Duration.

-> SomePace a
Pace.

Given an existentially-quantified distance and a duration, derives the pace.