Pacer.Configuration.Env.Types

Cached paths

data CachedPaths

Holds cached paths, for ensuring we do not make unnecessary calls (e.g. lookup the xdg config path twice).

Constructors

MkCachedPaths

Fields

currentDirectory :: Maybe (Path Abs Dir)

xdgConfigPath :: Maybe (Path Abs Dir)

Functions

getCachedXdgConfigPath :: forall (es :: [Effect]). (HasCallStack, PathReader :> es, State CachedPaths :> es) => Eff es (Path Abs Dir)

Retrieves the xdg config path, updating the cache if necessary.

getCachedCurrentDirectory :: forall (es :: [Effect]). (HasCallStack, PathReader :> es, State CachedPaths :> es) => Eff es (Path Abs Dir)

Retrieves the current directory, updating the cache if necessary. As the current directory may change, this is intended to be the _initial_ current directory i.e. wherever the program is invoked.

Logging Env

data LogEnv

Logging environment.

Constructors

MkLogEnv

Fields

logLevel :: Maybe LogLevel

logNamespace :: Namespace

logVerbosity :: LogVerbosity