Pacer.Configuration.Utils

Custom types and parsers

data DistancePaceArgs a

Constructors

MkDistancePaceArgs

Fields

mSomeDistance :: Maybe (SomeDistance a)

Possible distance.

mSomePace :: Maybe (SomePace a)

Possible pace.

dpArgsParser :: (Fromℚ a, Ord a, Parser a, Semifield a, Show a) => Parser (DistancePaceArgs a)

data DistanceDurationPaceArgs a

Args for derive and scale. All args are optional, though we require exactly 2 and 1, respectively.

Constructors

MkDistanceDurationPaceArgs

Fields

mDuration :: Maybe (Duration a)

Possible duration.

mPaceOptUnits :: Maybe (PaceOptUnits a)

Possible pace.

mSomeDistance :: Maybe (SomeDistance a)

Possible distance.

ddpArgsParser :: (Fromℚ a, Ord a, Parser a, Semifield a, Show a) => String -> Parser (DistanceDurationPaceArgs a)

type PaceOptUnits a = Either (SomePace a) (Duration a)

Represents the pace. For some conversions, the units are optional, hence we only require the underlying Duration (a Pace is just a duration with a unit, after all).

It is arguably simpler to require a unit for pace, and make the user use --duration for when they don't want a unit (e.g. convert or scale). So why do we allow no units? Consider deriving duration:

derive --distance '10 km' --pace 5m30s

Requring the user to type '/km' on the pace is redundant (hence annoying). Futhermore, we cannot replace this with '--duration 5m30s', since that has an entirely different meaning. Therefore it makes sense to have the units optional here, so we might as well make the units optional wherever sensible.

paceOptUnitsParserHelp :: (Fromℚ a, Ord a, Semifield a, Show a) => String -> Parser (PaceOptUnits a)

Other Parsers

distanceUnitParser :: Parser DistanceUnit

someDistanceParser :: (AMonoid a, Fromℚ a, Ord a, Parser a, Show a) => Parser (SomeDistance a)

somePaceParser :: (Fromℚ a, Ord a, Semifield a, Show a) => Parser (SomePace a)

Readers

readOsPath :: ReadM OsPath

readPath :: Exception e => (OsPath -> Either e (Path b t)) -> ReadM (Path b t)

readParseable :: Parser a => ReadM a

Commands

mkCommand :: String -> Parser a -> InfoMod a -> Mod CommandFields a

mkCommandDesc :: String -> InfoMod a

mkCommandDescChunk :: Chunk Doc -> InfoMod a

Misc

mkHelp :: forall (f :: Type -> Type) a. String -> Mod f a

mkMultiHelp :: forall (f :: Type -> Type) a. [String] -> Mod f a