Pacer.Data.Distance

Description

Provide distance types.

Distance

newtype Distance (d :: DistanceUnit) a

Represents a numeric distance with units.

Constructors

MkDistance

Fields

unDistance :: Positive a

Aliases

type Meters a = Distance 'Meter a

Alias for 'Distance Meters'.

type Kilometers a = Distance 'Kilometer a

Alias for 'Distance Kilometers'.

type Miles a = Distance 'Mile a

Alias for 'Distance Miles'.

Functions

liftDist :: forall a (d :: DistanceUnit). (Positive a -> Positive a) -> Distance d a -> Distance d a

liftDist2 :: forall a (d :: DistanceUnit). (Positive a -> Positive a -> Positive a) -> Distance d a -> Distance d a -> Distance d a

forceUnit :: forall (d :: DistanceUnit) a (e :: DistanceUnit). Distance d a -> Distance e a

Changes the distance without doing any conversions. Internal tool intended for forcing a universally quantified value to be a specific distance.

Distances

marathon :: forall (d :: DistanceUnit) a. (AMonoid a, Fromℚ a, Ord a, Show a, SingI d) => Distance d a

Marathon.

halfMarathon :: forall (d :: DistanceUnit) a. (AMonoid a, Fromℚ a, Ord a, Show a, SingI d) => Distance d a

Half marathon.

k_10 :: forall (d :: DistanceUnit) a. (AMonoid a, Fromℚ a, Ord a, Show a, SingI d) => Distance d a

10 km.

k_5 :: forall (d :: DistanceUnit) a. (AMonoid a, Fromℚ a, Ord a, Show a, SingI d) => Distance d a

5 km.

Some Distance

data SomeDistance a where

Distance, existentially quantifying the units.

Constructors

MkSomeDistance :: forall (d :: DistanceUnit) a. Sing d -> Distance d a -> SomeDistance a

Functions

liftSomeDist :: (Fromℤ a, Ord a, Semifield a, Show a) => (forall (d :: DistanceUnit). Distance d a -> Distance d a) -> SomeDistance a -> SomeDistance a

liftSomeDist2 :: (Fromℤ a, Ord a, Semifield a, Show a) => (forall (d :: DistanceUnit). Distance d a -> Distance d a -> Distance d a) -> SomeDistance a -> SomeDistance a -> SomeDistance a

Units

data DistanceUnit

Distance unit.

Constructors

Meter

Kilometer

Mile

Hiding Distance

class HasDistance a where

Class that provides a unified interface for handling distance units.

Associated Types

type DistanceVal a

type HideDistance a

Methods

distanceUnitOf :: a -> DistanceUnit

Retrieves the type of distance unit.

distanceOf :: a -> DistanceVal a

hideDistance :: a -> HideDistance a

Hides the distance unit.