Pacer.Data.Pace

Pace

data Pace (d :: DistanceUnit) a where

Represents a duration per distance e.g. 4'30" per kilometer. We only allow the distance to be kilometers or miles. Meters are disallowed as they chance for mistakes are high (i.e. almost all paces per meter are zero since they are so low).

Constructors

MkPace :: forall (d :: DistanceUnit) a. PaceDistF d => Duration a -> Pace d a

type family PaceDistF (d :: DistanceUnit) where ...

Equations

PaceDistF 'Meter = Unsatisfiable ('Text PaceMetersErrMsg)

PaceDistF 'Kilometer = ()

PaceDistF 'Mile = ()

Creation

mkPace :: forall (d :: DistanceUnit) a. PaceDistF d => Duration a -> Pace d a

Creates a pace from a duration.

Elimination

unPace :: forall (d :: DistanceUnit) a. Pace d a -> Duration a

Eliminates a pace to the underlying duration.

SomePace

data SomePace a where

Pace, existentially quantifying the distance unit.

Constructors

MkSomePace :: forall (d :: DistanceUnit) a. PaceDistF d => Sing d -> Pace d a -> SomePace a

Creation

mkSomePace :: forall (d :: DistanceUnit) a. (PaceDistF d, SingI d) => Pace d a -> SomePace a

Hides the distance.

Elimination

unSomePace :: SomePace a -> Duration a

Exposes the underlying duration.