Pacer.Utils

JSON

Encoding

encodeMaybe :: ToJSON v => (Key, Maybe v) -> [Pair]

Encodes a Maybe to a possibly empty list.

encodeMonoid :: (Eq v, Monoid v, ToJSON v) => (Key, v) -> [Pair]

Encodes a Monoid to a possibly empty list.

encodeMaybes :: ToJSON v => [(Key, Maybe v)] -> [Pair]

Encodes Maybes to a possibly empty list.

encodePretty :: ToJSON a => a -> LazyByteString

Decoding

(.:?:) :: (FromJSON (Item l), IsList l) => Object -> Key -> Parser l infixl 9

Decodes a json list into an IsList. Like (.:?), omitted keys are fine and decode to an empty list.

failUnknownFields

Arguments

:: Key
Key label to improve error message.

-> [Key]
Known keys.

-> KeyMap Value
Actual keys.

-> Parser ()

Fails if there are any unknown fields in the object.

decodeJson :: FromJSON a => ByteString -> Result AesonE a

Decodes json(c).

data AesonE

We use this rather than aeson's AesonException for two reasons:

AesonException does not have an NFData instance, which we require for the benchmarks.

Optional path improves the error message.

Constructors

MkAesonE (Maybe OsPath) String

readDecodeJson :: forall a (es :: [Effect]). (FileReader :> es, FromJSON a, HasCallStack) => Path Abs File -> Eff es a

Decodes a json(c) file.

Seq

seqGroupBy :: (a -> a -> Bool) -> Seq a -> Seq (NESeq a)

neSeqGroupBy :: (a -> a -> Bool) -> NESeq a -> NESeq (NESeq a)

Optics

data SomeSetter p where

Existential wrapper for an optic setter with new object.

Constructors

MkSomeSetter :: forall p x k (ix :: IxList). Is k A_Setter => Optic k ix p p x x -> x -> SomeSetter p

setMany' :: Foldable f => f (SomeSetter p) -> p -> p

Targets some object using a list of optics and the new values. Equivalent to calling set' multiple times.

Misc

type PaceMetersErrMsg = "Meters are disallowed in Pace; use km or mi."