Pacer.Utils.FileSearch

High-level

newtype FileSearchStrategy (f :: Type -> Type) (es :: [Effect])

FileSearchStrategy represents a single strategy for finding a single "type" of file(s), e.g. searching the current directory for the chart-requests file, or searching xdg for multiple activities files.

The semigroup instance takes the first "non-empty", per the FromAlt instance. In practice empty is generally Nothing (Maybe, for a single file), or [] (List/Seq, for multiple results).

In other words, a strategy amount to searching for some file(s) in a singular location. After we have a "success" (which be one or more files), we stop.

Constructors

MkFileSearchStrategy

Fields

unFileSearchStrategy :: FileSearch -> Eff es (f (Path Abs File))

resolveFilePath

Arguments

:: forall f (es :: [Effect]). (FromAlt f, HasCallStack, LoggerNS :> es)

=> Text
Text description of what we are trying to find. Used for logging.

-> FileSearch
Expected file_name(s) e.g. activities file(s).

-> [FileSearchStrategy f es]
Search strategies.

-> Eff es (f (Path Abs File))

Attempts to resolve a file, based on the expected file names and search strategies. The strategies are tried in order, returning the first non-empty, per FromAlt.

Strategies

findFilePath

Arguments

:: forall f (es :: [Effect]). (HasCallStack, PathReader :> es, Traversable f)

=> f OsPath
Maybe file.

-> FileSearchStrategy f es

If the parameter is not empty, parses to an absolute file(s). Ignores the parameter to searchFiles.

findCurrentDirectoryPath :: forall (f :: Type -> Type) (es :: [Effect]). (FromAlt f, HasCallStack, Logger :> es, LoggerNS :> es, PathReader :> es, Reader LogEnv :> es, State CachedPaths :> es) => FileSearchStrategy f es

findDirectoryPath specialized to the current directory.

findDirectoryPath

Arguments

:: forall (f :: Type -> Type) (es :: [Effect]). (FromAlt f, HasCallStack, Logger :> es, LoggerNS :> es, PathReader :> es, Reader LogEnv :> es)

=> Maybe OsPath
Directory to maybe search.

-> FileSearchStrategy f es

Searches for the files in the given directory. Otherwise returns empty.

findXdgPath :: forall (f :: Type -> Type) (es :: [Effect]). (FromAlt f, HasCallStack, Logger :> es, LoggerNS :> es, PathReader :> es, Reader LogEnv :> es, State CachedPaths :> es) => FileSearchStrategy f es

Searches for the given file in the xdg directory.

Low-level

File discovery

data DirNotExistsStrategy

How to handle a directory not existing.

Constructors

DirNotExistsOk
Directory not required; inexistence OK.

DirNotExistsFail
Directory required; inexistence should error.

data FileSearch

Represents searching for a particular file.

Constructors

SearchFileAliases FileAliases
Search for a file by aliases.

SearchFileInfix (Path Rel File) (Set OsPath)
Search for a file by infix pattern, optionally filtering by known extensions. For example, searching for "activities" and {"json", "csv"} will find all files with (case-insensitive) infix "activities" and extensions "json" or "csv".

data FileAliases

Represents a single file for which we want to search. Holds multiple aliases for the case where the file can have multiple names. For instance, both chart-requests.json and chart-requests.jsonc are valid expected names for the chart-requests file, so we search for both, even though we want at most one.

Constructors

MkFileAliases

Fields

aliases :: NonEmpty (Path Rel File)

extensions :: Set OsPath

directorySearch

Arguments

:: forall f (es :: [Effect]). (FromAlt f, HasCallStack, Logger :> es, LoggerNS :> es, PathReader :> es, Reader LogEnv :> es)

=> FileSearch
File names to search.

-> DirNotExistsStrategy
How to handle the directory not existing.

-> Path Abs Dir
Data dir to search.

-> Eff es (f (Path Abs File))

Attempts to discover files based on FileSearch i.e. "expected" filenames. For each search file s, produces a search result r per searchFile, and combines the result via (<|>).

In other words, depending on choice of type variable f, can return multiple results ([]) or a single result (Maybe).

searchFilesToList :: FileSearch -> ([Path Rel File], Set OsPath)

Errors

newtype DirNotFoundE

Exception for a directory not existing.

Constructors

MkDirNotFoundE OsPath

newtype FileNotFoundE

General exception for when a file at an expected path does not exist. We would normally use IOException for this, except we want a custom type so that we can ignore callstacks. We do not want to ignore IOException since other errors may throw them, and those we want to know about.

Constructors

MkFileNotFoundE OsPath

Misc

data MatchResult

File match comparison.

Constructors

MkMatchResult

Fields

patternResult :: Bool

The result for pattern comparisons i.e. file names.

extensionResult :: Maybe Bool

The result for file extension comparisons.

patternLog :: Text

Log message for pattern result.

extensionLog :: Text

Log message for extension result.

satisfiesPattern

Arguments

:: (Text -> Text -> Bool)
Pattern comparison function. The LHS is for the pattern whereas the RHS is the discovered file i.e. Pattern -> File -> Bool. The comparison should be for file _names_ only i.e. ignore directories and extensions.

-> Set OsString
Extension set. If empty, we ignore the discovered file's extension.

-> Path Rel File
Pattern i.e. alias or infix.

-> OsPath
Discovered file.

-> MatchResult

Returns the result of comparing a discovered file against a given pattern and possible extension set. For extension comparisons, we consider the "final" extension and the "full" extension e.g. ".gz" and ".tar.gz" in "foo.tar.gz", respectively. Parameters extensions should include the initial dot.

osPathToNameExts :: OsPath -> (OsPath, OsString, OsString)

Transforms a path to (name, final_extension, full_extension).

  pathToNameExts pathto/foo.tar.gz === (foo, .gz, .tar.gz)