Pacer.Web

Documentation

ensureWebDirExists

Arguments

:: forall (es :: [Effect]). (HasCallStack, FileWriter :> es, Logger :> es, LoggerNS :> es, PathReader :> es, PathWriter :> es)

=> Path Abs Dir
Path to web path dir. This should be WPaths.getWebPath, though we take it as a parameter to save a function call.

-> Bool

-> Eff es ()

Writes the web directory if it does not exist.