Pacer.Command.Chart

Description

Chart functionality.

Top level command

handle :: forall (es :: [Effect]). (HasCallStack, FileReader :> es, FileWriter :> es, Logger :> es, LoggerNS :> es, PathReader :> es, PathWriter :> es, Reader LogEnv :> es, ServerEff :> es) => ChartParamsFinal -> Eff es ()

Handles chart command.

Low level functions

type ChartPaths = (Path Abs File, Maybe (Path Abs File), NESeq (Path Abs File))

Misc

createChartSeq :: forall (es :: [Effect]). (HasCallStack, FileReader :> es, Logger :> es, LoggerNS :> es, Reader LogEnv :> es) => ChartPaths -> Eff es Charts

Given file paths to activities and chart requests, generates a sequence of charts.

updateLabels

Arguments

:: Map Timestamp (NESet Label)
Timestamp -> labels map.

-> SomeActivities a
All activities.

-> (SomeActivities a, Map Timestamp (NESet Label))
(New activities, unmatching timestamps).

Updates all activities with labels from the given map m. Returns a (possibly empty) subset of m that was unmatched i.e. had no corresponding activity in SomeActivities.