Pacer.Exception

Exceptions

data ChartFileMissingE

Exception for when we are expecting a required chart file to exist, but none was found. This differs from FileNotFoundE in that the latter is caused by a failure to find an extant file at an exact location. This exception, on the other hand, is caused by looking for a desired file in several locations, but none was found.

Constructors

MkChartFileMissingE

Fields

cliDataDir :: Maybe OsPath

configDataDir :: Maybe OsPath

currentDir :: Path Abs Dir

expectedFiles :: [Path Rel File]

expectedExts :: Set OsPath

xdgDir :: Path Abs Dir

newtype CreateChartE

Exception during chart creation.

Constructors

CreateChartFilterEmpty Text

data GarminE

Garmin exceptions.

Constructors

GarminDecode String
Error decoding a garmin file.

GarminMeters
Error for specifing garmin unit as meters.

GarminOther String
Generic garmin error.

GarminUnitRequired
Error for unspecified garmin distance unit.

Commands

data CommandConvertE

Exception for CLI scale command.

Constructors

CommandConvertArgs0
Wanted 1 quantity, received 0.

CommandConvertArgs2
Wanted 1 quantity, received 2.

CommandConvertPaceMeters
Error for converting pace with out unit in meters, not allowed.

data CommandDeriveE

Exception for CLI derive command.

Constructors

CommandDeriveArgs0
Wanted 2 quantities, received 0.

CommandDeriveArgs1
Wanted 2 quantities, received 1.

CommandDeriveArgs3
Wanted 2 quantities, received 3.

CommandDeriveDurationUnit
Output units are not used with derive duration.

CommandDeriveNoPaceUnit
Deriving distance requires pace with units.

CommandDerivePaceMeters
Error for deriving pace with out unit in meters, not allowed.

data CommandScaleE

Exception for CLI scale command.

Constructors

CommandScaleArgs0
Wanted 1 quantity, received 0.

CommandScaleArgs2
Wanted 1 quantity, received 2.

CommandScaleArgs3
Wanted 1 quantity, received 3.

CommandScaleDurationUnit
Output units are not used with scale duration.

CommandScalePaceMeters
Error for scaling pace with out unit in meters, not allowed.

CommandScalePaceUnitNoUnit Text
Error for scaling pace with out unit when original unit not given, nonsensical. The text arg is the original pace without units, so we can produce an example of correct usage.